v1.0.5. Change some DEVICE and MODEL private double parameters to float, and it leads to reduction of POWER_SYSTEM_DATABASE and DYNAMIC_MODEL_DATABASE memory usage but won't improve system performance. Remove some range check for COMPLEX_SPARSE_MATRIX_CSPARSE, and performance is improved by about 5%. If compare to v1.0.4, the performance is improved by about 10%.

+++ code/header/device/fixed_shunt.h
@@ -49,6 +49,6 @@ class FIXED_SHUNT : public DEVICE
         BUS* busptr;
         string identifier;
         bool status;
-        complex<double> nominal_impedance_shunt_in_MVA;
+        complex<float> nominal_impedance_shunt_in_MVA;
 };
 #endif // FIXED_SHUNT_H
