+++ code/header/model/sg_models/compensator_model/COMP.h
@@ -45,7 +45,7 @@ class COMP: public COMPENSATOR_MODEL
     private:
         void copy_from_const_model(const COMP& model);
 
-        double Xe;
+        float Xe;
 };
 
 #endif // EXCITER_MODEL_H
