+++ code/header/model/wtg_models/wt_turbine_model/wt_turbine_model.h
@@ -66,7 +66,7 @@ class WT_TURBINE_MODEL : public WTG_MODEL
     private:
         void common_constructor();
 
-        double damping;
+        float damping;
 };
 
 #endif // WT_TURBINE_MODEL_H
